un-asynchify, made it work
