+++ src/app/components/statistics/statistics.component.html
@@ -17,11 +17,11 @@
         </mat-card>
     </mat-tab>
     <mat-tab label="peers">
-        <mat-card *ngIf="(stats.peers)">
-            <h2>{{ stats.peers.length }} Peers</h2>
+        <mat-card *ngIf="(peers)">
+            <h2>{{ peers.length }} Peers</h2>
             <div>
                 <mat-form-field>
-                    <input matInput (keyup)="stats.setFilterForPeers($event.target.value)" placeholder="Filter">
+                    <input matInput (keyup)="setFilterForPeers($event.target.value)" placeholder="Filter">
                 </mat-form-field>
 
                 <mat-form-field>
@@ -32,18 +32,18 @@
                 <i>Click on peer for more details.</i>
             </div>
             <mat-chip-list>
-                <div *ngFor="let peer of stats.filteredPeers">
-                    <mat-chip (click)="stats.selectedPeer = peer">{{getShortedMultiAddress(peer.id)}}
+                <div *ngFor="let peer of filteredPeers">
+                    <mat-chip (click)="selectedPeer = peer">{{getShortedMultiAddress(peer.id)}}
                     </mat-chip>
                 </div>
             </mat-chip-list>
-            <div *ngIf="(stats.selectedPeer)">
-                <div>ID: {{stats.selectedPeer.id}}</div>
+            <div *ngIf="(selectedPeer)">
+                <div>ID: {{selectedPeer.id}}</div>
                 <div>
                     <ul>
                         Announced as:
-                        {{stats.selectedPeer.addrs.length}}
-                        <li *ngFor="let addr of stats.selectedPeer.addrs">
+                        {{selectedPeer.addrs.length}}
+                        <li *ngFor="let addr of selectedPeer.addrs">
                             {{addr}}
                         </li>
                     </ul>
