+++ src/app/components/statistics/statistics.component.ts
@@ -10,18 +10,23 @@ import { ToastService } from '../../service/toast.service';
     templateUrl: './statistics.component.html'
 })
 export class StatisticsComponent {
-    addPeerInput = '';
-    public peerFilter = '';
-    public selectedPeer = null;
+    public addPeerInput = '';
+
+    public filterForPeers: string = null;
+    public peers: { id: string, addrs: string[] }[];
+    public filteredPeers: { id: string, addrs: string[] }[] = null;
+    public selectedPeer: { id: string, addrs: string[] };
 
     public searchHash = '';
     public findHash: {search: string, result: string}  = null;
 
 
-    constructor(public stats: StatisticsService, public ipfs: IpfsService, public toast: ToastService) { }
-
-    displayedColumnsBandWith: string[] = ['key', 'value'];
-    displayedColumnsPeers: string[] = ['peer'];
+    constructor(public stats: StatisticsService, public ipfs: IpfsService, public toast: ToastService) {
+        stats.peersSubject.subscribe(f => {
+            this.peers = f;
+            this.filterPeers();
+        });
+     }
 
     public async lookupHash(hash: string) {
         this.findHash = {
@@ -40,5 +45,39 @@ export class StatisticsComponent {
         const length = 6;
         const shorted = multiaddr.slice(0, length) + '...' + multiaddr.slice(multiaddr.length - length, multiaddr.length);
         return shorted;
+    }
+
+    public async setFilterForPeers(filter: string) {
+        this.filterForPeers = filter;
+        await this.filterPeers();
+    }
+
+    private async filterPeers() {
+        if (!this.filterForPeers) {
+            this.filteredPeers = this.peers;
+        } else {
+            this.filteredPeers = this.peers.filter(r => {
+                if (r.id.toLowerCase().includes(this.filterForPeers.toLowerCase())) {
+                    return true;
+                }
+                for (let i = 0; i < r.addrs.length; i++) {
+                    const toCheck = r.addrs[i];
+                    console.log(toCheck);
+                    if (toCheck.toLowerCase().includes(this.filterForPeers.toLowerCase())) {
+                        return true;
+                    }
+                }
+                return false;
+            });
+            if (this.selectedPeer) {
+                if (!this.filteredPeers.find(f => f.id === this.selectedPeer.id)) {
+                    this.selectedPeer = null;
+                }
+            }
+            if (this.filteredPeers.length === 1) {
+                this.selectedPeer = this.filteredPeers[0];
+            }
+        }
+
     }
 }
